pyproject.toml
[project]
name = "mcp-server-uv"
version = "0.1.0"
description = "A Model Context Protocol server for uv package manager"
readme = "README.md"
authors = [
    { name = "liuheng", email = "liuheng@bonree.com" }
]
requires-python = ">=3.11"
dependencies = [
    "mcp-python>=0.1.0",
    "uv>=0.1.10"
]

[project.scripts]
mcp-server-uv = "mcp_server_uv:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=mcp_server_uv --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
